deferred feature extraction to hooks
